Add team page and optimize mentor data
In this commit, I've added a meet the team page and created an array of objects to hold teammember data.

Eventually, this will include all participants who would like to be included on the page, their LinkedIn and Twitter accounts, and a tag with their cohort dates.

This is abaultely a WIP, but to start, I've just got an image grid with the mentors pictures.