Build cohort & mentors section for full screen
There are a ton of styling tweaks that will need to be made here, but the content is organized in the correct section now. Small wins, LOL!

I've also made the .container class more specific should we need containers for other things in the future.

We will need to confirm who wants to be on this page, but I wanted to get as much in place as possible for progress sake. We can remove folks if necessary and add them when they confirm.

Here is a WIP screenshot of mentors: https://zappy.zapier.com/2ed9854dd067edb0af89e4bb0a541815.png
Here is a WIP screenshot of cohorts: https://zappy.zapier.com/07385233cb02534fb88abe125bd33a58.png