Add cohort 3 members
This commit adds photos and links for all of cohort 3 (some of these images may need to be standardized, but let's wait for the members to send us new ones if they want). It also adds a linkedin URL for ethan and a new photo for Claudia. Claudia's photo may need to be standardized as well.