Hug participants to midline on med screen... again
I thought there were bigger issues here and really second guessed my solution. But upon further investigation, it turned out I just needed to recenter the participant grid items for the small screen size.

So, this commit moves all the participants images to a cluster in the center when on a medium sized screen.

After researching, it seemed like there was no way to target a column on the grid and apply styles. Instead, I needed to target each element and apply the style to the grid item. Here are some more deets: https://stackoverflow.com/questions/46308048/how-to-target-a-specific-column-or-row-in-css-grid-layout#answer-46308148

I feel like it might look best if we also do something with the headings (like center them maybe?).

Screenshot of current state: https://zappy.zapier.com/84bafdf1facbb247286c48fedaaafb45.png