Add Lisa as PoC on CoC
I've added that folks can reach Lisa at our Coc email address and on Slack. I've removed the ability to DM me on Twitter. Originally I had that there so folks who weren't in Slack had a channel to reach me, but now that we have the email address, that requirement is satisfied.