Add Michelle as PoC for CoC
Michelle Sauque has agreed to join the Code of Conduct team, so I've added her to the Code of Contact details in our repo and on our website.

TODO: Ideally we'd all have Slack handles without spaces in the name. There is absolutely no harm in keeping them as-is, but it does look a bit strange on the website.