Preconnect to domains to increase performance
This PR adds `<link rel="preconnect">` and (as a fallback for older browsers) `<link rel="dns-prefetch">` tags to the document `<head>` to warm up those connections earlier in the render lifecycle.

It also changes the order of a couple of the elements in the `<head>` (which is, honestly, a total micro-optimization). Finally, it adds an `async` attribute our single `<script>` tag so that other rendering can happen while it loads and executes.