Add "pro-tip" directly to home page code
We needed to decouple the "pro-tip" piece from the application block stored in GraphCMS as part of our content redesign because they will end up on separate pages. In the meantime, while the main branch of the site doesn't include the content shift, I have hard-coded it on the homepage so that we don't lose that content over the next few weeks.