Default project mentors section to open
I send people to our Volunteer page on the regular to tell them to apply to be a project mentor, but if they don’t know that that’s what we call it they sometimes miss opening the `<details>` element for that role, leaving them unable to find the application form. This change opens that first blurb by default making it easier for potential mentors to apply for the program.